+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_edit_single_line.dart
@@ -101,6 +101,15 @@ class MPTH2FileEditStateEditSingleLine extends MPTH2FileEditState
         }
       case LogicalKeyboardKey.keyL:
         keyProcessed = onKeyLDownEvent(event);
+      case LogicalKeyboardKey.keyS:
+        if (!isCtrlPressed &&
+            !isMetaPressed &&
+            !isAltPressed &&
+            !isShiftPressed) {
+          th2FileEditController.elementEditController
+              .toggleSelectedLinePointsSmoothOption();
+          keyProcessed = true;
+        }
     }
 
     if (keyProcessed) {